Extract generic-renderer logic from page into its own component
