+++ databrowser/modules/generic-renderer/components/GenericBrowser.vue
@@ -1,0 +1,233 @@
+<template>
+  <div>
+    <div class="py-2">
+      <h3>Select an API{{ apiLoading ? ' (Loading...)' : null }}</h3>
+      <Select class="select-blue" @change="apiChanges">
+        <option :key="null" :value="null">-- Select an API --</option>
+        <option
+          v-for="key in $store.getters['remoteapi/apiKeys']"
+          :key="key"
+          :value="key"
+        >
+          {{ $store.getters['remoteapi/apiDescriptionByKey'](key) }}
+        </option>
+      </Select>
+    </div>
+    <div class="py-2">
+      <h3>
+        {{ 'Select an Endpoint' }}
+      </h3>
+      <Select
+        class="select-blue"
+        :disabled="currentDocument == null"
+        @change="pathChanges"
+      >
+        <option :key="null" :value="null">-- Select an endpoint --</option>
+        <option v-for="item in currentDocumentPaths" :key="item" :value="item">
+          {{ item }}
+        </option>
+      </Select>
+    </div>
+
+    <div v-if="fetchError != null" class="py-3">
+      <Alert
+        title="Fetch error"
+        :content="fetchError"
+        @close="fetchError = null"
+      ></Alert>
+    </div>
+
+    <GenericDataRenderer
+      v-if="filteredData != null"
+      class="bg-gray-100"
+      :filtered-data="filteredData"
+      :render-config="currentOpenApiRenderConfig"
+      @paginationChanges="paginationChanges"
+    ></GenericDataRenderer>
+
+    <GenericFilterRenderer
+      v-if="openApiEndpointPathItem != null"
+      class="bg-gray-100"
+      :filter-parameters="filterParameters"
+      @filterChanges="filterChanges"
+    ></GenericFilterRenderer>
+  </div>
+</template>
+
+<script lang="ts">
+import Vue from 'vue';
+import { OpenAPIV3 } from 'openapi-types';
+import Alert from '~/components/global/Alert.vue';
+import Select from '~/components/global/Select.vue';
+import { FilterChanges } from '~/../web-components/databrowser-generic/src/generic/GenericFilter';
+import {
+  PaginationChanges,
+  PageableList,
+} from '~/../web-components/databrowser-generic/src/generic/GenericList';
+import { OpenApiState } from '~/store/remoteapi';
+import { genericRenderConfig } from '~/modules/generic-renderer/config/generic-render.config';
+import {
+  ListConfig,
+  ResourceConfig,
+} from '~/../web-components/databrowser-generic/src/renderer/config.model';
+import GenericDataRenderer from '~/modules/generic-renderer/components/GenericDataRenderer.vue';
+import GenericFilterRenderer from '~/modules/generic-renderer/components/GenericFilterRenderer.vue';
+
+const concatFilters = (values: string[]) =>
+  values != null ? values.join(',') : '';
+
+export default Vue.extend({
+  components: { Alert, GenericDataRenderer, GenericFilterRenderer, Select },
+  data() {
+    return {
+      currentApiKey: null as unknown as string,
+      fetchError: null as string | null,
+      filteredData: null as unknown as PageableList | null,
+      filterParameters: null as
+        | (OpenAPIV3.ReferenceObject | OpenAPIV3.ParameterObject)[]
+        | null
+        | undefined,
+      openApiEndpointPath: null as string | null,
+      openApiEndpointPathItem: null as
+        | OpenAPIV3.PathItemObject
+        | null
+        | undefined,
+    };
+  },
+  computed: {
+    apiLoading(): boolean {
+      return this.currentApi != null && this.currentApi?.loading;
+    },
+    currentApi(): OpenApiState | undefined {
+      return this.$store.getters['remoteapi/apiByKey'](this.currentApiKey);
+    },
+    currentDocument(): OpenAPIV3.Document | undefined {
+      return this.$store.getters['remoteapi/documentByKey'](this.currentApiKey);
+    },
+    currentDocumentPaths(): string[] {
+      return this.currentDocument == null || this.currentDocument.paths == null
+        ? []
+        : Object.keys(this.currentDocument.paths);
+    },
+    currentOpenApiRenderConfig(): ListConfig | ResourceConfig | undefined {
+      return genericRenderConfig[this.currentApiKey]?.[
+        this.openApiEndpointPath as string
+      ];
+    },
+  },
+  methods: {
+    apiChanges(event: Event) {
+      this.currentApiKey = (event.target as HTMLSelectElement).value;
+      this.$store.dispatch('remoteapi/selectApi', {
+        key: this.currentApiKey,
+      });
+    },
+    pathChanges(event: Event) {
+      // Return early if no current docuement is set or if the current document has no paths
+      if (this.currentDocument == null || this.currentDocument.paths == null) {
+        this.openApiEndpointPath = null;
+        this.openApiEndpointPathItem = null;
+        this.filterParameters = null;
+        this.fetchError = null;
+        return;
+      }
+
+      this.openApiEndpointPath = (event.target as HTMLSelectElement).value;
+      this.openApiEndpointPathItem =
+        this.currentDocument.paths[this.openApiEndpointPath];
+
+      const deepCopy = JSON.parse(
+        JSON.stringify(this.openApiEndpointPathItem)
+      ) as OpenAPIV3.PathItemObject<number>;
+
+      this.filterParameters = deepCopy.get?.parameters;
+
+      // Reset previous data
+      this.filteredData = null;
+      // Reset previous error information
+      this.fetchError = null;
+    },
+    async fetchData(url: string) {
+      try {
+        this.fetchError = null;
+        const responseData = await this.$axios.$get(url);
+
+        // Check response type
+        if (responseData == null) {
+          // If respone is null or undefined, throw an error
+          throw new Error(`Response data from is empty, url was ${url}`);
+        } else if (responseData.TotalItems != null) {
+          // If respone contains a TotalItems property, we suspect that the result is
+          // a pageable result from Open Data Hub Tourim API
+          this.filteredData = responseData;
+        } else if (responseData instanceof Array) {
+          // If the result is an array, we suspect that the result is a non-pageable list.
+          // In that case, set the pagination data to being one page
+          this.filteredData = {
+            TotalResults: responseData.length,
+            TotalPages: 1,
+            CurrentPage: 1,
+            PreviousPage: null,
+            NextPage: null,
+            Items: responseData,
+          };
+        } else if (responseData instanceof Object) {
+          // The response may be an object (e.g. the response for a single entity)
+          // In that case, just use the respone data as it is.
+          this.filteredData = responseData;
+        }
+      } catch (e) {
+        // eslint-disable-next-line no-console
+        console.error(`Error while fetching ${url}`, e);
+        this.filteredData = null;
+        this.fetchError = e instanceof Error ? e.message : (e as string);
+      }
+    },
+    async filterChanges(event: CustomEvent<FilterChanges>) {
+      let path = this.openApiEndpointPath ?? '';
+
+      // Replace path params in URL
+      const pathFilters = event.detail.filter?.path;
+      if (pathFilters != null) {
+        Object.keys(pathFilters).forEach((filterName) => {
+          path = path.replaceAll(
+            `{${filterName}}`,
+            concatFilters(pathFilters[filterName])
+          );
+        });
+      }
+
+      // Add query params to URL
+      const queryFilters = event.detail.filter?.query;
+      if (queryFilters != null) {
+        const queryParams = Object.keys(queryFilters).reduce(
+          (prev, filterName) => [
+            ...prev,
+            `${filterName}=${concatFilters(queryFilters[filterName])}`,
+          ],
+          [] as string[]
+        );
+        if (queryParams.length) {
+          path += `?${queryParams.join('&')}`;
+        }
+      }
+
+      if (
+        this.currentDocument?.servers != null &&
+        this.currentDocument.servers.length > 0
+      ) {
+        // Take first server URL from OpenAPI document. Maybe this should be configurable
+        const fullUrl = this.currentDocument.servers[0].url + path;
+        await this.fetchData(fullUrl);
+      } else {
+        this.fetchError =
+          'No server URL found for current API (see OpenAPI document)';
+      }
+    },
+    async paginationChanges(event: CustomEvent<PaginationChanges>) {
+      const url = event.detail.url;
+      await this.fetchData(url);
+    },
+  },
+});
+</script>
